refactor: attempt at making pages server-rendered
